refactor(positioning): replace double for loop with numpy
broadcasting over the combination of our own 9x3 matrix of points in a
unit circle around us and the current position of our teammates.

It might make sense to increase the distance to filter by a factor, the
way it currently works a next position will be ignored, if from that
position a teammate is closer to the specific position and thus could
potentially get there faster.

This however ignores our orientation currently and that it is
preferrable to walk forward instead of backward.